Create README.md
Credit card fraud detection is a critical issue in the financial industry. The goal of this project is to develop a machine learning model that can accurately identify fraudulent transactions based on historical transaction data.

Dataset
The dataset used in this project is a commonly used credit card fraud detection dataset, which contains transactions made by European cardholders in September 2013. It contains only numerical input variables which are the result of a PCA transformation. Features V1, V2, ..., V28 are the principal components obtained with PCA. The only features which have not been transformed with PCA are 'Time' and 'Amount'. The feature 'Time' contains the seconds elapsed between this transaction and the first transaction in the dataset. The feature 'Amount' is the transaction Amount. The feature 'Class' is the response variable and it takes value 1 in case of fraud and 0 otherwise.